Made the workaround for Nlog.config more protected. Should not execute on a ISH.12 CD GH-73

+++ Source/Server/Install/Install-ISHDeployment.ps1
@@ -348,12 +348,15 @@ $installScriptBlock={
 
     #region Workaround GH-73
     $nlogConfigPath=Join-Path $CDPath "__InstallTool\nlog.config"
-    if(-not(Test-Path -Path ($nlogConfigPath+".bak")))
+    if(Test-Path -Path $nlogConfigPath)
     {
-        $null=Copy-Item -Path $nlogConfigPath -Destination ($nlogConfigPath+".bak")
-        $nlog=Get-Content -Path $nlogConfigPath -Raw
-        $nlog=$nlog.Replace('fileName="${serviceName}.${processId}.${creationDateTime}.log"','fileName="${processname:fullName=True}.${processId}.${creationDateTime}.log"')
-        [System.IO.File]::WriteAllText($nlogConfigPath,$nlog)
+        if(-not(Test-Path -Path ($nlogConfigPath+".bak")))
+        {
+            $null=Copy-Item -Path $nlogConfigPath -Destination ($nlogConfigPath+".bak")
+            $nlog=Get-Content -Path $nlogConfigPath -Raw
+            $nlog=$nlog.Replace('fileName="${serviceName}.${processId}.${creationDateTime}.log"','fileName="${processname:fullName=True}.${processId}.${creationDateTime}.log"')
+            [System.IO.File]::WriteAllText($nlogConfigPath,$nlog)
+        }
     }
     #endregion
